QueueDAG moved to top-level
