+++ examples/code/queue_dag.py
@@ -1,6 +1,5 @@
-import webber.queue as queue
 import time
-from webber import Promise
+import webber
 
 def writer(x):
     fd = open('myfile.txt', 'a+')
@@ -8,9 +7,9 @@ def writer(x):
     fd.close()
 
 
-dag = queue.QueueDAG()
+dag = webber.QueueDAG()
 x = dag.add_node(lambda: "1", iterator=100)
-y = dag.add_node(writer, Promise(x))
+y = dag.add_node(writer, webber.Promise(x))
 dag.add_edge(x, y)
 t = time.time()
 dag.execute(print_exc=True)
